Refactor completed
