+++ mm2src/coins/lightning/ln_utils.rs
@@ -1,30 +1,22 @@
 use super::*;
-use crate::lightning::ln_conf::{LightningCoinConf, LightningProtocolConf};
-use crate::lightning::ln_p2p::{connect_to_nodes_loop, init_peer_manager, ln_node_announcement_loop};
 use crate::lightning::ln_platform::{get_best_header, ln_best_block_update_loop, update_best_block};
 use crate::utxo::rpc_clients::BestBlock as RpcBestBlock;
-use crate::utxo::utxo_standard::UtxoStandardCoin;
-use crate::DerivationMethod;
 use bitcoin::hash_types::BlockHash;
 use bitcoin_hashes::{sha256d, Hash};
 use common::executor::{spawn, Timer};
 use common::log;
 use common::log::LogState;
 use common::mm_ctx::MmArc;
-use lightning::chain::keysinterface::{InMemorySigner, KeysInterface, KeysManager};
-use lightning::chain::{chainmonitor, Access, BestBlock, Watch};
+use lightning::chain::keysinterface::{InMemorySigner, KeysManager};
+use lightning::chain::{chainmonitor, BestBlock, Watch};
 use lightning::ln::channelmanager;
 use lightning::ln::channelmanager::{ChainParameters, ChannelManagerReadArgs, SimpleArcChannelManager};
-use lightning::routing::network_graph::{NetGraphMsgHandler, NetworkGraph};
+use lightning::routing::network_graph::NetworkGraph;
 use lightning::routing::scoring::Scorer;
+use lightning::util::config::UserConfig;
 use lightning::util::ser::ReadableArgs;
-use lightning_background_processor::BackgroundProcessor;
-use lightning_invoice::payment;
-use lightning_invoice::utils::DefaultRouter;
 use lightning_persister::storage::{FileSystemStorage, NodesAddressesMap};
 use lightning_persister::LightningPersister;
-use parking_lot::Mutex as PaMutex;
-use std::collections::HashMap;
 use std::fs::File;
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
@@ -44,29 +36,9 @@ pub type ChainMonitor = chainmonitor::ChainMonitor<
 
 pub type ChannelManager = SimpleArcChannelManager<ChainMonitor, Platform, Platform, LogState>;
 
-type Router = DefaultRouter<Arc<NetworkGraph>, Arc<LogState>>;
+fn ln_data_dir(ctx: &MmArc, ticker: &str) -> PathBuf { ctx.dbdir().join("LIGHTNING").join(ticker) }
 
-pub type InvoicePayer<E> = payment::InvoicePayer<Arc<ChannelManager>, Router, Arc<Mutex<Scorer>>, Arc<LogState>, E>;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct LightningParams {
-    // The listening port for the p2p LN node
-    pub listening_port: u16,
-    // Printable human-readable string to describe this node to other users.
-    pub node_name: [u8; 32],
-    // Node's RGB color. This is used for showing the node in a network graph with the desired color.
-    pub node_color: [u8; 3],
-    // Invoice Payer is initialized while starting the lightning node, and it requires the number of payment retries that
-    // it should do before considering a payment failed or partially failed. If not provided the number of retries will be 5
-    // as this is a good default value.
-    pub payment_retries: Option<usize>,
-    // Node's backup path for channels and other data that requires backup.
-    pub backup_path: Option<String>,
-}
-
-pub fn ln_data_dir(ctx: &MmArc, ticker: &str) -> PathBuf { ctx.dbdir().join("LIGHTNING").join(ticker) }
-
-pub fn ln_data_backup_dir(ctx: &MmArc, path: Option<String>, ticker: &str) -> Option<PathBuf> {
+fn ln_data_backup_dir(ctx: &MmArc, path: Option<String>, ticker: &str) -> Option<PathBuf> {
     path.map(|p| {
         PathBuf::from(&p)
             .join(&hex::encode(&**ctx.rmd160()))
@@ -75,7 +47,7 @@ pub fn ln_data_backup_dir(ctx: &MmArc, path: Option<String>, ticker: &str) -> Op
     })
 }
 
-async fn init_persister(
+pub async fn init_persister(
     ctx: &MmArc,
     ticker: String,
     backup_path: Option<String>,
@@ -98,7 +70,7 @@ async fn init_persister(
     Ok(persister)
 }
 
-fn init_keys_manager(ctx: &MmArc) -> EnableLightningResult<Arc<KeysManager>> {
+pub fn init_keys_manager(ctx: &MmArc) -> EnableLightningResult<Arc<KeysManager>> {
     // The current time is used to derive random numbers from the seed where required, to ensure all random generation is unique across restarts.
     let seed: [u8; 32] = ctx.secp256k1_key_pair().private().secret.into();
     let cur = SystemTime::now()
@@ -108,12 +80,12 @@ fn init_keys_manager(ctx: &MmArc) -> EnableLightningResult<Arc<KeysManager>> {
     Ok(Arc::new(KeysManager::new(&seed, cur.as_secs(), cur.subsec_nanos())))
 }
 
-async fn init_channel_manager(
+pub async fn init_channel_manager(
     platform: Arc<Platform>,
     logger: Arc<LogState>,
     persister: Arc<LightningPersister>,
     keys_manager: Arc<KeysManager>,
-    conf: LightningCoinConf,
+    user_config: UserConfig,
 ) -> EnableLightningResult<(Arc<ChainMonitor>, Arc<ChannelManager>)> {
     // Initialize the FeeEstimator. UtxoStandardCoin implements the FeeEstimator trait, so it'll act as our fee estimator.
     let fee_estimator = platform.clone();
@@ -156,7 +128,6 @@ async fn init_channel_manager(
         sha256d::Hash::from_slice(&best_block.hash.0).map_to_mm(|e| EnableLightningError::HashError(e.to_string()))?,
     );
     let (channel_manager_blockhash, channel_manager) = {
-        let user_config = conf.into();
         if let Ok(mut f) = File::open(persister.manager_path()) {
             let mut channel_monitor_mut_references = Vec::new();
             for (_, channel_monitor) in channelmonitors.iter_mut() {
@@ -233,7 +204,7 @@ async fn init_channel_manager(
     Ok((chain_monitor, channel_manager))
 }
 
-async fn persist_network_graph_loop(persister: Arc<LightningPersister>, network_graph: Arc<NetworkGraph>) {
+pub async fn persist_network_graph_loop(persister: Arc<LightningPersister>, network_graph: Arc<NetworkGraph>) {
     loop {
         if let Err(e) = persister.save_network_graph(network_graph.clone()).await {
             log::warn!(
@@ -245,7 +216,7 @@ async fn persist_network_graph_loop(persister: Arc<LightningPersister>, network_
     }
 }
 
-async fn persist_scorer_loop(persister: Arc<LightningPersister>, scorer: Arc<Mutex<Scorer>>) {
+pub async fn persist_scorer_loop(persister: Arc<LightningPersister>, scorer: Arc<Mutex<Scorer>>) {
     loop {
         if let Err(e) = persister.save_scorer(scorer.clone()).await {
             log::warn!(
@@ -257,7 +228,7 @@ async fn persist_scorer_loop(persister: Arc<LightningPersister>, scorer: Arc<Mut
     }
 }
 
-async fn get_open_channels_nodes_addresses(
+pub async fn get_open_channels_nodes_addresses(
     persister: Arc<LightningPersister>,
     channel_manager: Arc<ChannelManager>,
 ) -> EnableLightningResult<NodesAddressesMap> {
@@ -270,141 +241,4 @@ async fn get_open_channels_nodes_addresses(
             .any(|node_id| node_id == *pubkey)
     });
     Ok(nodes_addresses)
-}
-
-pub async fn start_lightning(
-    ctx: &MmArc,
-    platform_coin: UtxoStandardCoin,
-    protocol_conf: LightningProtocolConf,
-    conf: LightningCoinConf,
-    params: LightningParams,
-) -> EnableLightningResult<LightningCoin> {
-    // Todo: add support for Hardware wallets for funding transactions and spending spendable outputs (channel closing transactions)
-    if let DerivationMethod::HDWallet(_) = platform_coin.as_ref().derivation_method {
-        return MmError::err(EnableLightningError::UnsupportedMode(
-            "'start_lightning'".into(),
-            "iguana".into(),
-        ));
-    }
-
-    let platform = Arc::new(Platform::new(
-        platform_coin.clone(),
-        protocol_conf.network.clone(),
-        protocol_conf.confirmations,
-    ));
-
-    // Initialize the Logger
-    let logger = ctx.log.0.clone();
-
-    // Initialize Persister
-    let persister = init_persister(ctx, conf.ticker.clone(), params.backup_path).await?;
-
-    // Initialize the KeysManager
-    let keys_manager = init_keys_manager(ctx)?;
-
-    // Initialize the NetGraphMsgHandler. This is used for providing routes to send payments over
-    let network_graph = Arc::new(persister.get_network_graph(protocol_conf.network.into()).await?);
-    spawn(persist_network_graph_loop(persister.clone(), network_graph.clone()));
-    let network_gossip = Arc::new(NetGraphMsgHandler::new(
-        network_graph.clone(),
-        None::<Arc<dyn Access + Send + Sync>>,
-        logger.clone(),
-    ));
-
-    // Initialize the ChannelManager
-    let (chain_monitor, channel_manager) = init_channel_manager(
-        platform.clone(),
-        logger.clone(),
-        persister.clone(),
-        keys_manager.clone(),
-        conf.clone(),
-    )
-    .await?;
-
-    // Initialize the PeerManager
-    let peer_manager = init_peer_manager(
-        ctx.clone(),
-        params.listening_port,
-        channel_manager.clone(),
-        network_gossip.clone(),
-        keys_manager.get_node_secret(),
-        logger.clone(),
-    )
-    .await?;
-
-    let inbound_payments = Arc::new(PaMutex::new(HashMap::new()));
-    let outbound_payments = Arc::new(PaMutex::new(HashMap::new()));
-
-    // Initialize the event handler
-    let event_handler = Arc::new(ln_events::LightningEventHandler::new(
-        // It's safe to use unwrap here for now until implementing Native Client for Lightning
-        platform.clone(),
-        channel_manager.clone(),
-        keys_manager.clone(),
-        inbound_payments.clone(),
-        outbound_payments.clone(),
-    ));
-
-    // Initialize routing Scorer
-    let scorer = Arc::new(Mutex::new(persister.get_scorer().await?));
-    spawn(persist_scorer_loop(persister.clone(), scorer.clone()));
-
-    // Create InvoicePayer
-    let router = DefaultRouter::new(network_graph, logger.clone());
-    let invoice_payer = Arc::new(InvoicePayer::new(
-        channel_manager.clone(),
-        router,
-        scorer,
-        logger.clone(),
-        event_handler,
-        payment::RetryAttempts(params.payment_retries.unwrap_or(5)),
-    ));
-
-    // Persist ChannelManager
-    // Note: if the ChannelManager is not persisted properly to disk, there is risk of channels force closing the next time LN starts up
-    let channel_manager_persister = persister.clone();
-    let persist_channel_manager_callback =
-        move |node: &ChannelManager| channel_manager_persister.persist_manager(&*node);
-
-    // Start Background Processing. Runs tasks periodically in the background to keep LN node operational.
-    // InvoicePayer will act as our event handler as it handles some of the payments related events before
-    // delegating it to LightningEventHandler.
-    let background_processor = Arc::new(BackgroundProcessor::start(
-        persist_channel_manager_callback,
-        invoice_payer.clone(),
-        chain_monitor.clone(),
-        channel_manager.clone(),
-        Some(network_gossip),
-        peer_manager.clone(),
-        logger,
-    ));
-
-    // If channel_nodes_data file exists, read channels nodes data from disk and reconnect to channel nodes/peers if possible.
-    let open_channels_nodes = Arc::new(PaMutex::new(
-        get_open_channels_nodes_addresses(persister.clone(), channel_manager.clone()).await?,
-    ));
-    spawn(connect_to_nodes_loop(open_channels_nodes.clone(), peer_manager.clone()));
-
-    // Broadcast Node Announcement
-    spawn(ln_node_announcement_loop(
-        channel_manager.clone(),
-        params.node_name,
-        params.node_color,
-        params.listening_port,
-    ));
-
-    Ok(LightningCoin {
-        platform,
-        conf,
-        peer_manager,
-        background_processor,
-        channel_manager,
-        chain_monitor,
-        keys_manager,
-        invoice_payer,
-        persister,
-        inbound_payments,
-        outbound_payments,
-        open_channels_nodes,
-    })
 }
